Made snapshot and mode provider to App
