+++ src/Components/App.test.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import renderer from 'react-test-renderer';
+import { configure } from 'enzyme'
+import Adapter from 'enzyme-adapter-react-16'
+
+import App from './App';
+
+configure({ adapter: new Adapter() })
+
+it("Render App component", function () {
+  const AppComponent = renderer.create(
+    <App />
+  );
+  expect(AppComponent).toMatchSnapshot();
+});
